Initial migration from pack-it-forms
The Outpost integration functionality that was previously integrated
with the main pack-it-forms project is being split into it own
repository. This commit contains the initial version of the migrated
files modified so that they are expected to work in the new structure.

This has not seen any testing yet, so some fixup commits are expected
before it will be usable.