Ensure required files actually get copied
In the earlier version of the distribution script a few files weren't
copied into the distribution that should have been. Furthermore, when
the files that were removed to the integration repo the directory that
was the target for some of them no longer existed, which prevented the
copy from working.