Simplify installation instructions
These simplifications are intended to be more procedural and require
less knowledge and decision-making by the installer. They also
incorporate a recommendation to copy the PacFORMS version of
pac-read.exe into the pack-it-forms exec directory that was made by
Jim Clark based on his experience installing for less experienced
users.